feat: enable notebooks as module federation module
