+++ packages/notebooks/upstream/config/dotenv.js
@@ -158,7 +158,7 @@ const setupDotenvFilesForEnv = ({ env }) => {
   const DEPLOYMENT_MODE = process.env.DEPLOYMENT_MODE || 'kubeflow';
   const AUTH_METHOD = process.env.AUTH_METHOD || 'internal';
   const IMAGES_DIRNAME = process.env.IMAGES_DIRNAME || 'images';
-  const PUBLIC_PATH = process.env.PUBLIC_PATH || '/workspaces';
+  const PUBLIC_PATH = process.env.PUBLIC_PATH || '/';
   const SRC_DIR = path.resolve(RELATIVE_DIRNAME, process.env.SRC_DIR || TS_BASE_URL || 'src');
   const COMMON_DIR = path.resolve(RELATIVE_DIRNAME, process.env.COMMON_DIR || '../common');
   const DIST_DIR = path.resolve(RELATIVE_DIRNAME, process.env.DIST_DIR || TS_OUT_DIR || 'dist');
