+++ packages/notebooks/upstream/config/moduleFederation.js
@@ -1,3 +1,4 @@
+const path = require('path');
 const { ModuleFederationPlugin } = require('@module-federation/enhanced/webpack');
 
 const deps = require('../package.json').dependencies;
@@ -5,12 +6,11 @@ const deps = require('../package.json').dependencies;
 const moduleFederationConfig = {
   name: 'notebooks',
   filename: 'remoteEntry.js',
-
   shared: {
-    react: { singleton: true, eager: true, requiredVersion: deps.react },
-    'react-dom': { singleton: true, eager: true, requiredVersion: deps['react-dom'] },
-    'react-router': { singleton: true, eager: true, requiredVersion: deps['react-router'] },
-    'react-router-dom': { singleton: true, eager: true, requiredVersion: deps['react-router-dom'] },
+    react: { singleton: true, requiredVersion: deps.react },
+    'react-dom': { singleton: true, requiredVersion: deps['react-dom'] },
+    'react-router': { singleton: true, requiredVersion: deps['react-router'] },
+    'react-router-dom': { singleton: true, requiredVersion: deps['react-router-dom'] },
     '@patternfly/react-core': { singleton: true, requiredVersion: deps['@patternfly/react-core'] },
     '@openshift/dynamic-plugin-sdk': {
       singleton: true,
@@ -22,15 +22,11 @@ const moduleFederationConfig = {
     },
   },
   exposes: {
-    // TODO expose api. eg:
-    // './index': './src/plugin/index.tsx',
-    // './plugin': './src/plugin/index.tsx',
     './extensions': './src/odh/extensions',
   },
   // For module federation to work when optimization.runtimeChunk="single":
   // See https://github.com/webpack/webpack/issues/18810
   runtime: false,
-  // TODO generate types when exposing api
   dts: true,
 };
