+++ packages/notebooks/upstream/config/webpack.dev.js
@@ -76,6 +76,7 @@ module.exports = smp.wrap(
       optimization: {
         runtimeChunk: 'single',
         removeEmptyChunks: true,
+        splitChunks: false, // Disable automatic code splitting in dev mode
       },
       devServer: {
         host: HOST,
